(#6856) Copy dangling symlinks with 'links => manage' File resource.
When copying symlinks with 'links => manage', there is no need to
examine the content of the link target, since the link is supposed to be
copied as-is.  Skip copying this value to the resource from the metadata
(source), since the content will be nil in the dangling symlink case.

Paired-with: Jacob Helwig